% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_ref.R
\name{a_ref}
\alias{a_ref}
\title{Internal Data: Expected age per school level}
\format{
A data frame with elements specific to internal calculations.
}
\description{
\code{a_ref} is used within functions for specific calculations in the package.
This dataframe is not exported and is intended for internal use only.
}
\keyword{internal}
