% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rows2.R
\name{add_rows2}
\alias{add_rows2}
\title{Add rows to labelled data set.}
\usage{
add_rows2(..., id = NULL)
}
\arguments{
\item{...}{Other parameters to parse to function.}

\item{id}{Default NULL}
}
\value{
SAS labelled dataframe
}
\description{
Almost identical to the original \code{dplyr::add_row} but also looks for \code{format.sas} attribute which \code{haven} provides when loading as SAS dataset.
}
\details{
Created by PV (2023).
}
