% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aducust_flag.R
\name{aducust_flag}
\alias{aducust_flag}
\title{Custodial data flagging function}
\usage{
aducust_flag(
  data,
  dobmap,
  carer_map,
  flag_name = "carer_aducust",
  child_id_var = "rootnum",
  carer_id_var = "carer_rootnum",
  data_start_date = "ReceptionDate",
  data_end_date = "DischargeDate",
  dobmap_dob_var = "dob",
  child_start_age = 0,
  child_end_age = 18,
  carer_summary = FALSE,
  any_carer_summary = FALSE
)
}
\arguments{
\item{data}{Input dataset (carer aducust).}

\item{dobmap}{DOBmap file at the child level.}

\item{carer_map}{Mapping file with columns "child ID", "carer ID". Can have multiple rows per child (e.g., one per carer 1, carer 2, NEWBMID).}

\item{flag_name}{Name of flagging variable to return. Default \code{"carer_aducust"}.}

\item{child_id_var}{Variable denoting "child ID". Must exist and be called the same thing in \code{dobmap} and \code{carer_map}. Default \code{"rootnum"}.}

\item{carer_id_var}{Variable denoting "carer ID". Must exist in \code{carer_map}. Default \code{"carer_rootnum"}.}

\item{data_start_date}{Start date to consider in \code{data}. Corresponds to aducust \emph{start} date. Default \code{"ReceptionDate"}.}

\item{data_end_date}{End date to consider in \code{data}. Corresponds to aducust \emph{end} date. Default \code{"DischargeDate"}.}

\item{dobmap_dob_var}{Date of birth (DOB) variable in \code{dobmap}. Default \code{"dob"}.}

\item{child_start_age}{Numeric. Start (minimum) age (years) to consider for flagging (default \code{0}).}

\item{child_end_age}{Numeric. End (maximum) age (years) to consider for flagging (default \code{18}).}

\item{carer_summary}{Collapse aducust flags \emph{within carer} (i.e., for each \code{carer_id_var}). Default \code{FALSE}.}

\item{any_carer_summary}{Collapse aducust flags \emph{across carers}. Default \code{FALSE}.}
}
\value{
Flagged dataframe.
}
\description{
This function serves to flag whether a custodial record exists when a child is of a certain stage.
}
\details{
For more details, see the \href{https://dempz.github.io/WAACHShelp/articles/aducust_flag.html}{vignette}.
}
\note{
\itemize{
\item If \code{carer_summary} is \code{TRUE}, then we are flagging whether a specific carer has any aducust record.
\item Therefore, we are assessing whether a specific carer has any aducust records.
\item If \code{any_carer_summary} is \code{TRUE}, then we are flagging whether any carer (if multiple) have any aducust records.
\item If \code{any_carer_summary} is \code{TRUE}, \code{carer_summary} will be ignored.
}
}
\examples{
\dontrun{
# Example 1: Basic use
aducust_flag(data = carer_aducust \%>\% rename(carer_rootnum = root),
             dobmap = dobmap,
             carer_map = child_carer_map,
             child_id_var = "NEWUID",
             carer_id_var = "carer_rootnum",
             carer_summary = FALSE,
             any_carer_summary = TRUE
)
}
}
