% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_att_n.R
\name{check_att}
\alias{check_att}
\title{Education flagging function}
\usage{
check_att(
  dob,
  visibility_min = 2008,
  visibility_max = 2014,
  show_expectation = TRUE
)
}
\arguments{
\item{dob}{Date of individual. Must be class "Date".}

\item{visibility_min}{Minimum year in visibility window. Default \code{2008}.}

\item{visibility_max}{Maximum year in visibility window. Default \code{2014}.}

\item{show_expectation}{Logical. Show flagged dataframe with expected flag per year in visibility window. Default \code{TRUE}.}
}
\value{
Flagged dataframe.
}
\description{
This function serves to flag whether an individual is expected to have education data for every year within a "visibility window" based on their date of birth.
The fact that "staggered" (June/July) years were introduced in Western Australia in 1997 has been built into the function.
}
\examples{
check_att(dob = as.Date("2000-05-30"))
}
