% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_classify_list.R
\name{colname_classify_broad}
\alias{colname_classify_broad}
\title{Internal Data List: Broad morbidity column classification}
\format{
A list with elements specific to internal calculations.
}
\description{
\code{colname_classify_broad} is used within functions for specific calculations in the package.
This list is not exported and is intended for internal use only.
}
\keyword{internal}
