% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colname_classify_list.R
\name{colname_classify_specific}
\alias{colname_classify_specific}
\title{Internal Data List: Specific morbidity column classification}
\format{
A list with elements specific to internal calculations.
}
\description{
\code{colname_classify_specific} provides additional data needed for calculations in the package.
This list is also for internal use and is not exported.
}
\keyword{internal}
