% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_markdown.R
\name{create_markdown}
\alias{create_markdown}
\title{Create Template for Quarto Markdown}
\usage{
create_markdown(file_name = NULL, directory = "reports", ext_name = "html")
}
\arguments{
\item{file_name}{A string. The name of the new Quarto markdown (.qmd) file. This must be provided.}

\item{directory}{A string. The name of the directory to plate the files. Default is NULL. Requires user specification}

\item{ext_name}{A string. The extension type to create. Default "html" (alternatives: "word").}
}
\description{
This function creates a HTML or Word template in the current working directory
using a specified Quarto extension. It copies the template files to the
\verb{_extensions/} directory and generates a new Quarto markdown (.qmd) file.
}
\details{
Adapted from \link[thekidsbiostats:create_template]{create_template} from \url{https://github.com/The-Kids-Biostats/thekidsbiostats}.

The function first checks whether a \verb{_extensions/} directory exists in the current working
directory. If not, it creates one. It then copies the necessary extension files from the
package's internal data to the \verb{_extensions/} directory. Finally, it creates
a new Quarto markdown file based on the extension template.

By default, the \code{reports} folder will be selected to house the report.

For more details, see the \href{../doc/create_markdown.html}{vignette}.
}
\note{
The function assumes that the package \code{WAACHShelp} contains the necessary extension files
under \verb{ext_qmd/_extensions/}.
}
\examples{
\dontrun{
create_markdown(file_name = "my_doc", ext_name = "word")
}

}
