% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-02-icd_flagging.R
\name{icd_flagging}
\alias{icd_flagging}
\title{Helper Function #3}
\usage{
icd_flagging(
  data,
  flag_category,
  icd_list,
  flag_other_varname,
  icd_dat,
  colname_classify_specific
)
}
\description{
This is an internal helper function.
}
\keyword{internal}
