% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-04.1-icd_consolidated_v2.R
\name{icd_morb_flag}
\alias{icd_morb_flag}
\title{ICD flagging function}
\usage{
icd_morb_flag(
  data,
  dobmap,
  flag_category,
  flag_other_vals,
  flag_other_varname,
  under_age = FALSE,
  age = 18,
  person_summary = FALSE
)
}
\arguments{
\item{data}{Input dataset (morbidity).}

\item{dobmap}{DOBMap file corresponding to input dataset.}

\item{flag_category}{Type of flag to generate. Takes values from reference file (e.g., MH_morb, Sub_morb, etc.) or "Other" for custom ICD specification and flagging.}

\item{flag_other_vals}{Flag variable ICD values (specified only when \code{flag_category == "Other"}). Input as list with keys "diag_ediag" (for searching across diagnosis, ediag variables) and "ecode" (for searching across ecode variables).}

\item{flag_other_varname}{Flag variable name (specified only When \code{flag_category == "Other"}). Input as character string.}

\item{under_age}{Return additional variables corresponding to when participant was strictly under \code{age} y.o.. Uses DOBMap DOB and subadm morbidity admission date. Variables have suffix "_under{age}".}

\item{age}{Integer. Age to consider for the \code{under_age} variable (default 18).}

\item{person_summary}{Summarise results at a person-level.}
}
\value{
Flagged dataframe.
}
\description{
This function serves to add flags to an input data set (morbidity at this stage) pursuant to ICD codes.
Flags can be added for general categories based on pre-established ICD codes (e.g., any mental health contact, any substance-related contact, any alcohol/tobacco-related contact etc.) or add a custom set of ICD codes.
The file with these pre-established ICD codes are saved as an .RData file and are trivial to change.
}
\examples{
# Example 1: Basic use
icd_morb_flag(data = morb,
              dobmap = dob,
              flag_category = "MH_morb" # Create any MH contact flag
              under_age = T, # Return additional set of flags depending on whether participant is under 18 at time of admission.
              age = 18
              )

# Example 2: Specify custom set of ICD codes to base flags on
icd_morb_flag(data = morb,
              dobmap = dob,
              flag_category = "Other",
              flag_other_varname = "testing_var",                                # Create new flagging variable called "testing_var"
              flag_other_vals = list("diag_ediag" = c("abc24.21", "xyz4231.2")), # Base flag on searching across the "diagnosis" and "ediag" variables (vector corresponds to ICD codes)
              under_age = T,
              age = 25,                                                          # Return additional set of flags depending on whether participant is under 25 at time of admission.
              person_level = F
              )
}
