% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/person_summary.R
\name{person_summary}
\alias{person_summary}
\title{Summarise a dataframe to a person level}
\usage{
person_summary(data, flag_category, flag_category_val = "Yes", grouping_var)
}
\arguments{
\item{data}{Input dataframe.}

\item{flag_category}{Name of the variable to perform classification on.}

\item{flag_category_val}{String value of the \code{flag_category} variable the flag will be judged against. Defaults to "Yes".}

\item{grouping_var}{Grouping ID variable that identifies potentially multiple records per participant.}
}
\description{
In cases where we have a \emph{long} data frame (i.e., multiple rows of data per participant) with a flag against each record (e.g., variable \code{x} = "Yes"/"No"), this function will collapse this dataframe to a participant level.
}
\details{
The collapsing is based on the value of \code{flag_category_val} and a grouping variable (participant ID) \code{grouping_var}.

Specifically, records will be collapsed such that:
\itemize{
\item If \emph{any} record(s) for a participant is equal to \code{flag_category_val}, then return "Yes".
\item If \emph{all} record(s) for a participant are not equal to \code{flag_category_val}, then return "No".
}
}
\examples{
\dontrun{
person_summary(data = dat,
               flag_category = "variable_x",
               flag_category_val = "Yes",
               grouping_var = "record_id"
               )
}

}
