% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_contents.R
\name{proc_contents}
\alias{proc_contents}
\title{Compile metadata from dataframe}
\usage{
proc_contents(df)
}
\arguments{
\item{df}{Dataframe to input.}
}
\value{
Dataframe
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function was deprecated because it was no longer required by analysts.
}
\details{
This little function pulls out the labels and formats from a dataframe and compiles this metadata as a dataframe.

Imitates the "proc contents" function of SAS.

Created by PV (2023).
}
\examples{
proc_contents(iris)

}
