% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptab.R
\name{ptab}
\alias{ptab}
\title{Display dataframe as HTML table}
\usage{
ptab(tt, shading = NULL, caption = " ", bodysize = 11)
}
\arguments{
\item{tt}{Dataframe intended to display as a html table.}

\item{shading}{Vector denoting the lines of the table that might want to be shaded. For example, to shade the first, third and fifth rows, supply \code{shading = c(1, 3, 5)}.}

\item{caption}{Caption to add to the data set.}

\item{bodysize}{Text size of the table's content.}
}
\value{
Two-way table
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function was deprecated because it was no longer required by analysts.
}
\details{
Created by PV (2023).
}
\keyword{internal}
