% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_waachs.R
\name{save_waachs}
\alias{save_waachs}
\title{Save dataframe}
\usage{
save_waachs(dataframe, path, filename)
}
\arguments{
\item{dataframe}{Input dataset to save.}

\item{path}{Path to save file to.}

\item{filename}{Name of the file to save.}
}
\value{
Saved object
}
\description{
General function to save a dataset in usable formats across different platforms. Specifically, will return .csv, .sas7bdat, and .RDS files.
}
