% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_table.R
\name{sum_tab}
\alias{sum_tab}
\title{Summary tabling function}
\usage{
sum_tab(data, ...)
}
\arguments{
\item{data}{Input dataset.}

\item{...}{Any other argument relevant to \code{gtsummary::tbl_summary}.}
}
\value{
Summary table.
}
\description{
General tabling function to standardise output and formatting across analysts and projects.
Based on gtsummary::tbl_summary
}
