% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_waachs.R
\name{theme_waachs}
\alias{theme_waachs}
\title{Apply WAACHS theme to ggplot2 plots}
\usage{
theme_waachs(
  base_size = 12,
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{Base font size}

\item{base_line_size}{Base line size (default \code{base_size/22})}

\item{base_rect_size}{Base rectangle size (default \code{base_size/22})}
}
\value{
A list of ggplot2 theme elements and scale adjustments.
}
\description{
This function applies a custom theme to ggplot2 plots, incorporating colours to align with the project's visual identity.
}
\details{
The function determines the operating system and selects appropriate font names for Windows or other systems. It also adjusts colour scales.
}
\examples{

ggplot(mtcars,
       aes(x = mpg, y = wt, col = factor(cyl))) +
  geom_point() +
  theme_waachs()

}
