% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpose_tblsum.R
\name{transpose_tblsum}
\alias{transpose_tblsum}
\title{WAACHS transpose \code{gtsummary::tbl_summary} table}
\usage{
transpose_tblsum(tbl, ...)
}
\arguments{
\item{tbl}{tbl_summary input (of class \code{c("tbl_summary", "gtsummary")}).}

\item{...}{Other parameters (not currently in use).}
}
\value{
data.frame with transposed summary table.
}
\description{
Quick helper function to easily transpose \code{gtsummary::tbl_summary} tables. Best used in conjunction with \code{WAACHShelp::waachs_table}.
}
\examples{
mtcars \%>\%
  mutate(cyl = as_factor(cyl)) \%>\%
  select(cyl, mpg, disp, hp, wt, am) \%>\%
  tbl_summary(by = cyl) \%>\%
  modify_header(label ~ "cyl") \%>\% # Re-label "Characteristic" by stratification variable ("cy")
  transpose_tblsum() \%>\%
  waachs_table()

}
