% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoway.R
\name{twoway}
\alias{twoway}
\title{HTML summary table}
\usage{
twoway(var1, var2, data = NULL, var2lab = NULL)
}
\arguments{
\item{var1}{Vector of first variable}

\item{var2}{Vector of second variable}

\item{data}{Dataset containing \code{var1} and \code{var2}}

\item{var2lab}{Label for \code{var2}.}
}
\value{
Two-way table
}
\description{
Two-way table similar to the "proc freq" function of SAS with two variables
}
\details{
Created by PV (2023).
}
