% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/val_filt.R
\name{val_filt}
\alias{val_filt}
\title{Basic ICD code function}
\usage{
val_filt(input_vec, letter, lower, upper)
}
\arguments{
\item{input_vec}{Vector of all admissible ICD codes to filter on.}

\item{letter}{Letter to base filtration on (if purely numeric use empty string "")}

\item{lower}{Lower bound on the numeric element of the ICD code (includes numerics >=lower).}

\item{upper}{Upper bound on the numeric element of the ICD code (includes numerics <=upper).}
}
\value{
Vector with filtered ICD codes.
}
\description{
This function is already used by the team, and filters alphanumeric ICD-9 and ICD-10 codes pursuant to requirements.
}
\examples{
val_filt(val3, "F", 10.0, 10.9) # Filter ICD codes in val3 to those between F10 and F10.9 (inclusive).

}
