% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waachs_palette.R
\name{waachs_palette}
\alias{waachs_palette}
\title{WAACHS colour palette}
\usage{
waachs_palette(
  type = "discrete",
  n,
  visualisation = F,
  bias = 2,
  interpolate = "spline",
  ...
)
}
\arguments{
\item{type}{Type of colour palette to render (values "discrete", "continuous").}

\item{n}{Number of colours to generate in palette (if \code{type == "continuous"}).}

\item{visualisation}{(Default \code{TRUE}). Return visualisation of the colour spectrum to aid in decision making.}

\item{bias}{A positive nmber representing the spacing between colours at the high end. Parsed to \code{grDevices::colorRamp} (default 2).}

\item{interpolate}{Interpolation algorithm to parse to \code{grDevices::colorRamp} (default "spline").}

\item{...}{Miscellaneous arguments to parse to \code{grDevices::colorRampPalette}.}
}
\value{
Vector with colour hex codes. If \code{type == "continuous"} returns vector of length \code{n}. If \code{visualisation == TRUE} returns a list containing colour palette vector and \code{ggplot2} visualisation of colour spectrum.
}
\description{
Function to return discrete or continuous colour palette based on WAACHS logo colours.
}
\examples{
colours <- waachs_palette(type = "continuous",
                          n = 100,
                          visualisation = TRUE) # Render colour palette
print(colours$palette) # Print colour palette
print(colours$plot)    # Print plot

}
