% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waachs_table.R
\name{waachs_table}
\alias{waachs_table}
\title{WAACHS table formatting function.}
\usage{
waachs_table(
  x,
  font.size = 8,
  font.size.header = 10,
  line.spacing = 1.5,
  padding = 2,
  body_bg_col = "#FEF0D8",
  header_bg_col = "#89A1AD",
  header_text_col = "black",
  font.family = "Barlow"
)
}
\arguments{
\item{x}{A table, typically a data.frame, tibble, or output from gtsummary.}

\item{font.size}{The font size for text in the body of the table, defaults to 8 (passed throught to set_flextable_defaults).}

\item{font.size.header}{The font size for text in the header of the table, defaults to 10.}

\item{line.spacing}{Line spacing for the table, defaults to 1.5 (passed throught to set_flextable_defaults).}

\item{padding}{Padding around all four sides of the text within the cell, defaults to 2 (passed throught to set_flextable_defaults).}

\item{body_bg_col}{Body background colour (default WAACHS cream).}

\item{header_bg_col}{Header background colour (default WAACHS blue).}

\item{header_text_col}{Header text colour (default black).}

\item{font.family}{Font family for plot (default Barlow).}
}
\description{
Function to apply consistent formatting to summary tables rendered using R. Works with functions from the \code{gtsummary} package, or dataframes.
}
\details{
Inspired and based on \code{thekidsbiostats::thekids_table()}.
}
\examples{
head(mtcars) \%>\%
  waachs_table()

}
