% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rows2.R
\name{add_rows2}
\alias{add_rows2}
\title{Add rows to labelled data set.}
\usage{
add_rows2(..., id = NULL)
}
\arguments{
\item{...}{Other parameters to parse to function.}

\item{id}{Default NULL}
}
\value{
SAS labelled dataframe
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function was deprecated because it was no longer required by analysts.
}
\details{
Almost identical to the original \code{dplyr::add_row} but also looks for \code{format.sas} attribute which \code{haven} provides when loading as SAS dataset.

Created by PV (2023).
}
