% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-01-icd_extraction.R
\name{icd_extraction}
\alias{icd_extraction}
\title{Helper Function #2}
\usage{
icd_extraction(
  data,
  flag_category,
  diag_type,
  diag_type_custom_vars = NULL,
  diag_type_custom_params,
  icd_dat,
  colname_classify_specific
)
}
\description{
This is an internal helper function.
}
\keyword{internal}
