% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-proc_freq.R
\name{proc_freq}
\alias{proc_freq}
\title{HTML summary table}
\usage{
proc_freq(var1, data = NULL, sort = NULL, min.frq = 0)
}
\arguments{
\item{var1}{Name of the variable}

\item{data}{Dataset that contains \code{var1}}

\item{sort}{Optional argument which can take on "asc" or "desc" to indicate the type of sort required.}

\item{min.frq}{Minimum frequency}
}
\value{
Dataframe
}
\description{
Renders a HTML table ready for plonking into a HTML or word document.
}
\details{
Renders output similar to the "proc freq" function of SAS.

Created by PV (2023).
}
\examples{
test = iris
proc_freq(Species, test)
test[1:4, "Species"] <- NA
proc_freq(Species, test)

}
