% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumfun.R
\name{sumfun}
\alias{sumfun}
\title{Summary function reminiscent of Stata's "tabset".}
\usage{
sumfun(x, na.rm = TRUE, ...)
}
\arguments{
\item{x}{Vector from input dataframe to summarise.}

\item{na.rm}{(Default TRUE) to remove NA (missing) observations from summary.}

\item{...}{Any other arguments parsed into component base R functions.}
}
\value{
Summary table with n (length), miss (number of missing observations), mean, sd (standard deviation), med (median), q25 (first quartile), q75 (third quartile), min (minimum value), max (maximum value).
}
\description{
Created by PV (2023).
}
\examples{
sumfun(iris$Sepal.Length)
}
