% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoway.R
\name{twoway}
\alias{twoway}
\title{HTML summary table}
\usage{
twoway(var1, var2, data = NULL, var2lab = NULL)
}
\arguments{
\item{var1}{Vector of first variable}

\item{var2}{Vector of second variable}

\item{data}{Dataset containing \code{var1} and \code{var2}}

\item{var2lab}{Label for \code{var2}.}
}
\value{
Two-way table
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function was deprecated because it was no longer required by analysts.
}
\details{
Two-way table similar to the "proc freq" function of SAS with two variables

Created by PV (2023).
}
